back/requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
fuzzywuzzy==0.18.0
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-Levenshtein-wheels==0.13.2
python3-openid==3.2.0
pytz==2023.3.post1
rapidfuzz==3.5.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
sqlparse==0.4.4
thefuzz==0.20.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
